Move block controls into main top bar
