+++ js/app/packages/block-image/component/HeaderControls.tsx
@@ -7,14 +7,10 @@ import {
   BlockItemSplitLabel,
   SplitPermissionsBadge,
 } from '@app/component/split-layout/components/SplitLabel';
-import {
-  SplitToolbarLeft,
-  SplitToolbarRight,
-} from '@app/component/split-layout/components/SplitToolbar';
 import { useBlockId } from '@core/block';
 import { ReferencesModal } from '@core/component/ReferencesModal';
 import { ShareButton } from '@core/component/TopBar/ShareButton';
-import { blockMetadataSignal } from '@core/signal/load';
+import { blockFileSignal, blockMetadataSignal } from '@core/signal/load';
 import { useGetPermissions } from '@core/signal/permissions';
 import {
   useBlockDocumentDownloadName,
@@ -23,24 +19,18 @@ import {
 import { downloadFile } from '@filesystem/download';
 import Download from '@icon/regular/download.svg';
 import { createCallback } from '@solid-primitives/rootless';
-import { toast } from 'core/component/Toast/Toast';
-import { useGetFileBlob } from '../signal/blockData';
 
-export function TopBar() {
+export function HeaderControls() {
   const blockId = useBlockId();
+  const imageFile = blockFileSignal.get;
   const name = useBlockDocumentName();
   const downloadName = useBlockDocumentDownloadName();
-  const getBlob = useGetFileBlob();
   const userPermissions = useGetPermissions();
 
   const downloadDocument = createCallback(async () => {
-    try {
-      const blob = await getBlob();
-      downloadFile(blob, downloadName());
-    } catch (e) {
-      console.error('error downloading file', e);
-      toast.failure('Error downloading file');
-    }
+    const file = imageFile();
+    if (!file) return;
+    downloadFile(file, downloadName());
   });
 
   const ops: FileOperation[] = [
@@ -60,37 +50,33 @@ export function TopBar() {
   return (
     <>
       <SplitHeaderLeft>
-        <BlockItemSplitLabel />
-      </SplitHeaderLeft>
-      <SplitToolbarLeft>
-        <div class="p-1">
-          <SplitFileMenu
-            id={blockId}
-            itemType="document"
-            name={name()}
-            ops={ops}
-          />
-        </div>
-      </SplitToolbarLeft>
-      <SplitToolbarRight>
-        <div class="flex items-center p-1">
-          <ReferencesModal
-            documentId={blockId}
-            documentName={name()}
-            buttonSize="sm"
-          />
-          <div class="flex items-center">
-            <SplitPermissionsBadge />
-            <ShareButton
+        <div class="flex items-center gap-2 min-w-0">
+          <BlockItemSplitLabel />
+          <div class="flex items-center gap-2 shrink-0">
+            <SplitFileMenu
               id={blockId}
-              name={name()}
-              userPermissions={userPermissions()}
               itemType="document"
-              owner={blockMetadataSignal()?.owner}
+              name={name()}
+              ops={ops}
             />
+            <ReferencesModal
+              documentId={blockId}
+              documentName={name()}
+              buttonSize="sm"
+            />
+            <div class="flex items-center">
+              <SplitPermissionsBadge />
+              <ShareButton
+                id={blockId}
+                name={name()}
+                userPermissions={userPermissions()}
+                itemType="document"
+                owner={blockMetadataSignal()?.owner}
+              />
+            </div>
           </div>
         </div>
-      </SplitToolbarRight>
+      </SplitHeaderLeft>
     </>
   );
 }
