+++ js/app/packages/block-pdf/component/HeaderControls.tsx
@@ -10,10 +10,6 @@ import {
   BlockItemSplitLabel,
   SplitPermissionsBadge,
 } from '@app/component/split-layout/components/SplitLabel';
-import {
-  SplitToolbarLeft,
-  SplitToolbarRight,
-} from '@app/component/split-layout/components/SplitToolbar';
 import { useHasModificationData } from '@block-pdf/signal/save';
 import { useHasComments } from '@block-pdf/store/comments/commentStore';
 import { doPrint } from '@block-pdf/util/printUtil';
@@ -45,7 +41,7 @@ import { LocationType, useCreateShareUrl } from '../signal/location';
 import { MarkupToolbar } from './MarkupToolbar';
 import { PageNumberInput } from './PageNumberInput';
 
-export function TopBar() {
+export function HeaderControls() {
   const isAuth = useIsAuthenticated();
   const documentId = useBlockId();
   const hasModificationData = useHasModificationData();
@@ -170,54 +166,52 @@ export function TopBar() {
   return (
     <>
       <SplitHeaderLeft>
-        <BlockItemSplitLabel />
+        <div class="flex items-center gap-2 min-w-0">
+          <BlockItemSplitLabel />
+          <div class="flex items-center gap-2 shrink-0">
+            <Show when={pdfDocumentProxy()}>
+              <div class="flex items-center">
+                <SplitFileMenu
+                  id={documentId}
+                  itemType="document"
+                  name={fileName()}
+                  ops={ops}
+                />
+                <div class="w-5" />
+                <PageNumberInput />
+                <div class="w-5" />
+                {ENABLE_PDF_MARKUP && <MarkupToolbar />}
+              </div>
+            </Show>
+            <Show when={ENABLE_REFERENCES_MODAL}>
+              <ReferencesModal
+                documentId={documentId}
+                documentName={fileName()}
+                buttonSize="sm"
+              />
+            </Show>
+            <DocumentPropertiesModal
+              documentId={documentId}
+              blockType="pdf"
+              buttonSize="sm"
+            />
+            <div class="flex items-center">
+              <SplitPermissionsBadge />
+              <ShareButton
+                id={documentId}
+                name={fileName()}
+                userPermissions={userPermissions()}
+                copyLink={copyLink}
+                itemType="document"
+                owner={blockMetadataSignal()?.owner}
+              />
+            </div>
+          </div>
+        </div>
       </SplitHeaderLeft>
       <SplitHeaderRight>
         <BlockLiveIndicators />
       </SplitHeaderRight>
-      <SplitToolbarLeft>
-        <Show when={pdfDocumentProxy()}>
-          <div class="flex items-center p-1">
-            <SplitFileMenu
-              id={documentId}
-              itemType="document"
-              name={fileName()}
-              ops={ops}
-            />
-            <div class="w-5" />
-            <PageNumberInput />
-            <div class="w-5" />
-            {ENABLE_PDF_MARKUP && <MarkupToolbar />}
-          </div>
-        </Show>
-      </SplitToolbarLeft>
-      <SplitToolbarRight>
-        <div class="flex items-center p-1">
-          <Show when={ENABLE_REFERENCES_MODAL}>
-            <ReferencesModal
-              documentId={documentId}
-              documentName={fileName()}
-              buttonSize="sm"
-            />
-          </Show>
-          <DocumentPropertiesModal
-            documentId={documentId}
-            blockType="pdf"
-            buttonSize="sm"
-          />
-          <div class="flex items-center">
-            <SplitPermissionsBadge />
-            <ShareButton
-              id={documentId}
-              name={fileName()}
-              userPermissions={userPermissions()}
-              copyLink={copyLink}
-              itemType="document"
-              owner={blockMetadataSignal()?.owner}
-            />
-          </div>
-        </div>
-      </SplitToolbarRight>
     </>
   );
 }
