+++ js/app/packages/block-unknown/component/HeaderControls.tsx
@@ -7,10 +7,6 @@ import {
   BlockItemSplitLabel,
   SplitPermissionsBadge,
 } from '@app/component/split-layout/components/SplitLabel';
-import {
-  SplitToolbarLeft,
-  SplitToolbarRight,
-} from '@app/component/split-layout/components/SplitToolbar';
 import { useBlockId } from '@core/block';
 import { ReferencesModal } from '@core/component/ReferencesModal';
 import { ShareButton } from '@core/component/TopBar/ShareButton';
@@ -26,7 +22,7 @@ import { createCallback } from '@solid-primitives/rootless';
 import { toast } from 'core/component/Toast/Toast';
 import { useGetFileBlob } from '../signal/blockData';
 
-export function TopBar() {
+export function HeaderControls() {
   const blockId = useBlockId();
   const fileName = useBlockDocumentName();
   const downloadName = useBlockDocumentDownloadName();
@@ -60,37 +56,33 @@ export function TopBar() {
   return (
     <>
       <SplitHeaderLeft>
-        <BlockItemSplitLabel />
-      </SplitHeaderLeft>
-      <SplitToolbarLeft>
-        <div class="p-1">
-          <SplitFileMenu
-            id={blockId}
-            itemType="document"
-            name={fileName()}
-            ops={ops}
-          />
-        </div>
-      </SplitToolbarLeft>
-      <SplitToolbarRight>
-        <div class="flex items-center p-1">
-          <ReferencesModal
-            documentId={blockId}
-            documentName={fileName()}
-            buttonSize="sm"
-          />
-          <div class="flex items-center">
-            <SplitPermissionsBadge />
-            <ShareButton
+        <div class="flex items-center gap-2 min-w-0">
+          <BlockItemSplitLabel />
+          <div class="flex items-center gap-2 shrink-0">
+            <SplitFileMenu
               id={blockId}
-              name={fileName()}
-              userPermissions={userPermissions()}
               itemType="document"
-              owner={blockMetadataSignal()?.owner}
+              name={fileName()}
+              ops={ops}
             />
+            <ReferencesModal
+              documentId={blockId}
+              documentName={fileName()}
+              buttonSize="sm"
+            />
+            <div class="flex items-center">
+              <SplitPermissionsBadge />
+              <ShareButton
+                id={blockId}
+                name={fileName()}
+                userPermissions={userPermissions()}
+                itemType="document"
+                owner={blockMetadataSignal()?.owner}
+              />
+            </div>
           </div>
         </div>
-      </SplitToolbarRight>
+      </SplitHeaderLeft>
     </>
   );
 }
